backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
# Redis será instalado por celery[redis]
# redis==5.0.1  # Comentado para evitar conflicto con celery[redis]
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Librerías de seguridad adicionales
cryptography==41.0.7
bcrypt==4.1.2
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0

# Validación y sanitización
bleach==6.1.0
validators==0.22.0

# Rate limiting y seguridad
slowapi==0.1.9
python-multipart==0.0.6

# Logging y monitoreo
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# RabbitMQ y mensajería asíncrona
pika==1.3.2
aio-pika==9.3.1
celery[redis]==5.3.4

# Meilisearch para búsqueda de texto completo
meilisearch==0.31.0
meilisearch-python-async==1.8.1

# Procesamiento asíncrono adicional
aiofiles==23.2.1

# Plantillas HTML
jinja2==3.1.2
